+++ src/components/containers/movie-info/index.js
