+++ src/components/containers/movie-info/movie-info.js
