+++ src/components/containers/movie-info/movie-info.test.js
