+++ src/components/pages/main-page/index.js
