+++ src/components/pages/main-page/main-page.test.js
