+++ src/components/pages/movie-page/movie-page.js
@@ -1,12 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import CardFilmBg from "../../components/card-film-bg";
-import Header from "../../components/header";
-import MovieCardWrap from "../../components/movie-card-wrap";
-import MovieInfo from "../../components/movie-info";
+import CardFilmBg from "../../containers/card-film-bg";
+import Header from "../../containers/header";
+import MovieCardWrap from "../../containers/movie-card-wrap";
+import MovieInfo from "../../containers/movie-info";
 import {MovieMainWrapper, PageContentWrapper} from "../../wappers";
-import CardFilmList from "../../components/card-film-list";
+import CardFilmList from "../../containers/card-film-list";
+import Footer from "../../containers/footer";
 
 const MoviePage = ({film, films, onCardClick}) => {
   return (
@@ -24,6 +25,7 @@ const MoviePage = ({film, films, onCardClick}) => {
           <h2 className="catalog__title">More like this</h2>
           <CardFilmList films={films} onCardClick={onCardClick}/>
         </section>
+        <Footer/>
       </PageContentWrapper>
     </React.Fragment>
   );
